refactor(vats): move extract, makePromiseSpace from boot to utils
